🐛 chore(conftest): refactor Spark session creation
Refactored the Spark session creation in the conftest to be more readable and consistent. In the previous implementation, there was an error as the chaining of methods was not correctly formatted. This change ensures that the creation of the Spark session is done in a clearer and more uniform manner, promoting code maintainability.