♻️ refactor(data_preprocessing): rename split_data_with_reserve to split_data
The function `split_data_with_reserve` in the data preprocessing module has been renamed to `split_data`. This change was made to simplify the function name and make it more intuitive, while still retaining its original functionality. All references of this function in the module have also been updated according to this change.